SOCINT-266 Use /survey endpoint for available fulfilments Moved RHSvc webform call and AIMS calls
