+++ app/utils.py
@@ -1,6 +1,5 @@
 import string
 import re
-import jwt
 
 from aiohttp.client_exceptions import (ClientResponseError)
 from .exceptions import InactiveCaseError, InvalidEqPayLoad, InvalidDataError, InvalidDataErrorWelsh, \
@@ -373,64 +372,6 @@ class FlashMessage:
         return json_return
 
 
-class AddressIndex(View):
-
-    @staticmethod
-    def generate_jwt(request):
-        key = request.app['ADDRESS_INDEX_SVC_KEY']
-        token = jwt.encode({}, key, algorithm="HS256")
-        return token
-
-    @staticmethod
-    async def get_postcode_return(request, postcode, display_region):
-        postcode_return = await AddressIndex.get_ai_postcode(request, postcode)
-
-        address_options = []
-
-        if display_region == 'cy':
-            cannot_find_text = 'Ni allaf ddod o hyd i fy nghyfeiriad'
-        else:
-            cannot_find_text = 'I cannot find my address'
-
-        for singleAddress in postcode_return['response']['addresses']:
-            address_options.append({
-                'value': singleAddress['uprn'],
-                'label': {
-                    'text': singleAddress['formattedAddress']
-                },
-                'id': singleAddress['uprn']
-            })
-
-        address_options.append({
-            'value': 'xxxx',
-            'label': {
-                'text': cannot_find_text
-            },
-            'id': 'xxxx'
-        })
-
-        address_content = {
-            'postcode': postcode,
-            'addresses': address_options,
-            'total_matches': postcode_return['response']['total']
-        }
-
-        return address_content
-
-    @staticmethod
-    async def get_ai_postcode(request, postcode):
-        ai_svc_url = request.app['ADDRESS_INDEX_SVC_URL']
-        ai_epoch = request.app['ADDRESS_INDEX_EPOCH']
-        token = AddressIndex.generate_jwt(request)
-        url = f'{ai_svc_url}/addresses/rh/postcode/{postcode}?limit=5000&epoch={ai_epoch}'
-        headers = {'Authorization': 'Bearer ' + token}
-        return await MakeRequest.make_request(request,
-                                              'GET',
-                                              url,
-                                              headers=headers,
-                                              return_json=True)
-
-
 class RHService(View):
     @staticmethod
     async def post_surveylaunched(request, case, adlocation):
@@ -448,21 +389,3 @@ class RHService(View):
                                               f'{rhsvc_url}/surveyLaunched',
                                               auth=request.app['RHSVC_AUTH'],
                                               request_json=launch_json)
-
-    @staticmethod
-    async def post_webform(request, form_data):
-        form_json = {
-            'category': form_data['category'],
-            'region': form_data['region'],
-            'language': form_data['language'],
-            'name': form_data['name'],
-            'description': form_data['description'],
-            'email': form_data['email'],
-            'clientIP': View.single_client_ip(request)
-        }
-        rhsvc_url = request.app['RHSVC_URL']
-        return await MakeRequest.make_request(request,
-                                              'POST',
-                                              f'{rhsvc_url}/webform',
-                                              auth=request.app['RHSVC_AUTH'],
-                                              request_json=form_json)
