+++ tests/unit/helpers.py
@@ -350,7 +350,7 @@ class TestHelpers(RHTestCase):
 
     async def check_post_enter_address(self, display_region):
         with self.assertLogs('respondent-home', 'INFO') as cm, \
-                mock.patch('app.utils.AddressIndex.get_ai_postcode') as mocked_get_ai_postcode:
+                mock.patch('app.comms.aims.Postcode.get_ai_postcode') as mocked_get_ai_postcode:
             mocked_get_ai_postcode.return_value = self.ai_postcode_results
 
             response = await self.client.request('POST',
@@ -389,7 +389,7 @@ class TestHelpers(RHTestCase):
 
     async def check_post_enter_address_input_returns_no_results(self, display_region):
         with self.assertLogs('respondent-home', 'INFO') as cm, \
-                mock.patch('app.utils.AddressIndex.get_ai_postcode') as mocked_get_ai_postcode:
+                mock.patch('app.comms.aims.Postcode.get_ai_postcode') as mocked_get_ai_postcode:
             mocked_get_ai_postcode.return_value = self.ai_postcode_no_results
 
             response = await self.client.request('POST',
@@ -450,7 +450,7 @@ class TestHelpers(RHTestCase):
 
     async def check_post_select_address_no_selection_made(self, display_region):
         with self.assertLogs('respondent-home', 'INFO') as cm, \
-                mock.patch('app.utils.AddressIndex.get_ai_postcode') as mocked_get_ai_postcode:
+                mock.patch('app.comms.aims.Postcode.get_ai_postcode') as mocked_get_ai_postcode:
             mocked_get_ai_postcode.return_value = self.ai_postcode_results
 
             response = await self.client.request('POST',
@@ -467,7 +467,7 @@ class TestHelpers(RHTestCase):
 
     async def check_post_select_address(self, display_region, region):
         with self.assertLogs('respondent-home', 'INFO') as cm, \
-                mock.patch('app.utils.AddressIndex.get_ai_postcode') as mocked_get_ai_postcode, mock.patch(
+                mock.patch('app.comms.aims.Postcode.get_ai_postcode') as mocked_get_ai_postcode, mock.patch(
                 'app.comms.rhsvc.Cases.get_cases_by_uprn') as mocked_get_case_by_uprn:
             mocked_get_ai_postcode.return_value = self.ai_postcode_results
             if region == 'W':
@@ -558,7 +558,7 @@ class TestHelpers(RHTestCase):
 
     async def check_post_confirm_address_input_invalid(self, display_region):
         with self.assertLogs('respondent-home', 'INFO') as cm, \
-                mock.patch('app.utils.AddressIndex.get_ai_postcode') as mocked_get_ai_postcode, mock.patch(
+                mock.patch('app.comms.aims.Postcode.get_ai_postcode') as mocked_get_ai_postcode, mock.patch(
                 'app.comms.rhsvc.Cases.get_cases_by_uprn') as mocked_get_case_by_uprn:
             mocked_get_ai_postcode.return_value = self.ai_postcode_results
             mocked_get_case_by_uprn.return_value = self.rhsvc_case_by_uprn_hh_e
@@ -574,7 +574,7 @@ class TestHelpers(RHTestCase):
 
     async def check_post_confirm_address_input_no_selection(self, display_region):
         with self.assertLogs('respondent-home', 'INFO') as cm, \
-                mock.patch('app.utils.AddressIndex.get_ai_postcode') as mocked_get_ai_postcode, mock.patch(
+                mock.patch('app.comms.aims.Postcode.get_ai_postcode') as mocked_get_ai_postcode, mock.patch(
                 'app.comms.rhsvc.Cases.get_cases_by_uprn') as mocked_get_case_by_uprn:
             mocked_get_ai_postcode.return_value = self.ai_postcode_results
             mocked_get_case_by_uprn.return_value = self.rhsvc_case_by_uprn_hh_e
@@ -590,7 +590,7 @@ class TestHelpers(RHTestCase):
 
     async def check_post_confirm_address_input_no(self, display_region):
         with self.assertLogs('respondent-home', 'INFO') as cm, \
-                mock.patch('app.utils.AddressIndex.get_ai_postcode') as mocked_get_ai_postcode:
+                mock.patch('app.comms.aims.Postcode.get_ai_postcode') as mocked_get_ai_postcode:
             mocked_get_ai_postcode.return_value = self.ai_postcode_results
 
             response = await self.client.request('POST',
